refactor: test environment
